refactoring login window
